app/src/main/java/com/android/streetworkapp/MainActivity.kt
package com.android.streetworkapp

import android.os.Bundle
import android.util.Log
import androidx.activity.ComponentActivity
import androidx.activity.compose.setContent
import androidx.compose.runtime.Composable
import androidx.navigation.compose.rememberNavController

class MainActivity : ComponentActivity() {
  override fun onCreate(savedInstanceState: Bundle?) {
    super.onCreate(savedInstanceState)
    Log.d("MainActivity", "Setup content")
    setContent { Streetworkapp() }
  }
}

@Composable
fun Streetworkapp() {
  val navController = rememberNavController()
}
